Add OOP WIP
