+++ src/main/java/FileException.java
@@ -1,0 +1,5 @@
+public class FileException extends Exception {
+    public FileException(String string) {
+        super(string);
+    }
+}
