+++ src/main/java/Storage.java
@@ -1,0 +1,76 @@
+import java.io.*;
+import java.util.ArrayList;
+
+public class Storage {
+    private String filePath;
+    private ArrayList<String> text = new ArrayList<>();
+    private ArrayList<Boolean> status = new ArrayList<>();
+    public Storage(String filePath) {
+        this.filePath = filePath;
+    }
+
+    public TwoArrayList load() throws FileException {
+        try {
+            File file = new File(filePath);
+            BufferedReader reader = new BufferedReader(new FileReader(file));
+            String line = reader.readLine();
+
+            while (line != null) {
+                text.add(Parser.FileInputToTask(line));
+                status.add(Parser.FileInputToTaskStatus(line));
+                line = reader.readLine();
+            }
+            reader.close();
+            return new TwoArrayList(text, status);
+
+//            PrintStream originalOut = System.out;
+//            System.setOut(new PrintStream(new OutputStream() {
+//                @Override
+//                public void write(int b) {
+//                    // Do nothing
+//                }
+//            }));
+//
+//            while (line != null) {
+//
+//                String[] splitString = line.split("\\|", 0);
+//                String taskStatus = splitString[1];
+//                String taskInput = line.replaceAll("\\| 0 \\|", "")
+//                        .replaceAll("\\| 1 \\|", "")
+//                        .replaceAll("\\|", "/")
+//                        .replaceAll("T", "todo")
+//                        .replaceAll("D", "deadline")
+//                        .replaceAll("E", "event")
+//                        .replaceAll("  ", " ");
+//
+//                addTask(taskInput);
+//                boolean isTaskDone = taskStatus.trim().equals("1") ? true : false;
+//                changeTaskStatus(counter, isTaskDone);
+//                counter++;
+//                System.setOut(originalOut);
+//
+//            }
+        } catch (FileNotFoundException e) {
+            throw new FileException("Oops, Arona cannot find your task file!");
+        } catch (IOException e) {
+            throw new FileException("Sensei! There were some errors reading tasks from the file: " + e.getMessage());
+        }
+    }
+
+    public void save(ArrayList<Task> tasks) throws FileException {
+        try {
+            File file = new File(filePath);
+            BufferedWriter writer = new BufferedWriter(new FileWriter(file));
+
+            for (Task task : tasks) {
+                // System.out.println(task.toString());
+                writer.write(Parser.taskToFileOutput(task));
+                writer.newLine();
+            }
+            writer.flush();
+            writer.close();
+        } catch (IOException e ) {
+            throw new FileException("Sensei! There were some errors reading tasks from the file: " + e.getMessage());
+        }
+    }
+}
