+++ src/main/java/Ui.java
@@ -1,0 +1,21 @@
+public class Ui {
+//    public void greetings() {
+//        String logo ="    _____ \n" +
+//                /* */                "   /  _  \\_______   ____   _____ _____ \n" +
+//                /*   */              "  /  /_\\  \\_  __  \\/  _ \\ /     \\___  \\ \n" +
+//                /*   */      " /    |    \\  | \\_ ( <_> )   |   \\/ __ \\_ \n" +
+//                /* */                " \\____|__  /__|    \\____/|___|_  (____  / \n" +
+//                /* */                "         \\/                    \\/     \\/ \n";
+//        System.out.println("開始中... \n" + logo);
+////        String reply = "Hi! I'm " + this.name + ". What can I do for you?";
+//        String reply = "こんにちは先生、私は" + this.name + "アロナです. \n"
+//                + "どういうご用件ですか?　\n"
+//                + "ここで先生のスケジュールが決まります";
+//        System.out.println(reply);
+//    }
+//
+//    public void quitApplication() {
+//        String reply = "Goodbye sensei! Hope to see you soon!";
+//        System.out.println(reply);
+//    }
+}
